resets the testnet with a new genesis block (#3848)
before mainnet many miners and mining pools were mining on the testnet. the
hashpower on the testnet at that time pushed the difficulty very high.

now that that mining power has migrated to mainnet the testnet chain is
effectively halted.

resetting the testnet with a new genesis block will allow us to get the testnet
running again with lower mining power.