adds a simulation for fetching transactions from getAccountTransaction
we have heard reports that the 'wallet/getAccountTransaction' endpoint does not
always return every transaction.

this simulation creates three nodes and has two 'sender' nodes repeatedly send
transactions to the 'receiver' node. for each transaction the simulation checks
that the transaction can be found on the sender node and the receiver node using
the 'wallet/getAccountTransaction' endpoint over HTTP.

running the simulation for ~1000 blocks did not find any transactions missing
from either the sender or receiver.