adds account flag to faucet service command (#3874)
the faucet service always uses the default account for transactions.

this prevents us from combining the api-node and pool-node because mining and
pool payouts also use the default account. this could create issues if the
faucet and pool are sending transactions out from the same account.

adds a flag to set the account to send faucet payouts from.