removes unnecessary note decryption in wallet transaction RPCs (#3872)
the wallet/getAccountTransaction and wallet/getAccountTransactions RPC endpoints
decrypt notes from the transaction so that we can return data for any output
notes that an account sent to another address in the transaction.

we have to do this decryption because the walletDb doesn't store decrypted notes
that an account doesn't own.

however, this decryption is applied to all notes in the transaction. we don't
need to decrypt any notes that we already store in decrypted form in the
walletdb (i.e., notes the account owns). we also don't need to try to decrypt
any other output notes if the account was not the sender of the transaction
since we won't be able to decrypt them.

avoids unnecessary decryption and adds unit tests.