fix(ironfish): Check for transactions with duplicate nullifiers earlier
While it is not possible to double spend in this way due to checking for
this when adding a block to the chain, this kind of transaction could
cause issues for miners by getting them to create invalid blocks. By
checking for this on gossip and when adding a transaction to the
mempool, we remove that possibility.

It is not necessary to check this in all of the places that we are, but
it is a relatively cheap check, so if we re-introduce the flaw in one
place in the future, it will still get caught.