fix(ironfish): add additional transaction validation (#3901)
* fix(ironfish): add additional transaction validation

This is a low-cost check for transactions now that we have a transaction
index for the blockchain. This would catch transaction replays in
scenarios where spends might not be present. For example, a mint when
the minimum fee is 0 is technically replayable. By adding this check, we
ensure that even if the fee gets set to 0, we are not vulnerable to such
bugs.

* update test to not re-use the same transaction

* refactor: add specific error, change order of checks

Changing the order of the checks was just to keep other tests from
needing refactors. Does the same job.