perf: refactor messages to reduce a duplicate serialization
The 3 message types that implement the `Serializable` interface have
essentially a 2-step serialization process. Step 1: serialize the
metadata/header, and step 2: serialize the payload data.

We were instantiating a new bufio writer and calling `.render()` in both
steps, despite step 1 calling step 2 and immediately writing the
combined serialized message into a buffer.

By changing the payload serialization to take in an already instantiated
bufio writer, we can write directly into the same writer, and not have
to `.render()` it twice.

This should lead to substantial performance gains for both the worker
pool and the peer network.