creates spend witnesses using confirmation range (#3918)
supports creating spends with a confirmation range. if a spend is created with
the chain head at some block A and the chain reorgs to block B, then the spends
created at block A will be invalid because the commitment won't match the root
hash of the notes tree at that note size.

using the note tree size from a confirmed block instead of the current chain
head will reduce the likelihood that a reorg will render a spend invalid.