checks if submitted block is heavier when chain changed
when a miner submits a block to the mining manager the manager discards the
block if the chain has changed. that is, if the submitted block no longer
connects to the chain head.

however, if the submitted block is heavier than the curremt chain head then the
manager should add this block to the chain.

- removes unnecessary check that chain head exists
- does not discard submitted block if it is heavier than the current chain head