sets work on submitted block before checking isBlockHeavier
the 'work' on a block header is only set in 'chain.connect', so checking
'isBlockHeavier' in 'submitBlockTemplate' will always return false unless we set
the work first.

only sets the work and checks if the block is heavier if the chain has changed.
this is to avoid duplicating this work if possible and avoid looking up the
previous block if possible