removes loops from getNewBlockTransactions
when creating a new block template the mining manager collects a list of
transactions to include in the block.

the mining manager iterates over transactions in the mempool, verifies the
spends in the transactions, and tracks the total size and total fee from
included transactions.

we currently iterate over the list of included transactions twice after
collecting them: once to read the fee from the transaction and again to sum the
fees.

these changes remove the extra iteration over transactions for fees.
performance gains will likely be neglible.