regenerates fixtures and fixes outdated wallet tests
changes in #3830 inadvertently fixed a bug whereby transactions that did not
send any iron (e.g., mint transactions with 0 fee) would still include a spend
to cover the fee of 0.

this bug created fixtures for several tests in which fee was 0, but a spend was
included anyway.

after regenerating fixtures following the fix, these tests failed.

- do not expect transaction assetBalanceDeltas to include the native asset when
  fee is 0 and no iron is sent or received (i.e., mint or burn with 0 fee)
- fixes getTransactionType to check whether the account was the sender on any
  notes in case the transaction does not include spends from the account