Allow the sapling params to be downloaded at build time
Currently the `ironfish-rust/src/sapling_params` directory contains
almost 90 MB of files (which contain random data and cannot be
compressed). To reduce the size of the `ironfish` crate, this commit
adds build logic that allows that directory to optionally contain no
param files, and fetches them from GitHub.

In particular:
- files are fetched from the GitHub `master` branch if and only if they
  are not present in the `ironfish-rust/src/sapling_params` directory;
- files are checked for integrity using SHA-512, to detect incorrect or
  tampered downloads;
- when performing multiple builds, the files are downloaded and verified
  only during the first build, and results are cached for subsequent
  runs.