handles errors sending jobs to worker threads (#3984)
* handles errors sending jobs to worker threads

the worker pool tracks the state of a job by assigning it a status and by
assigning it to a worker. that assignment, placing the job in the worker's jobs
map, occurs before sending the job to the worker. the job's status is also set
to 'executing' before sending the job to the worker thread.

if an error occurs in sending the job to the worker, such as when serializing
the job, then we do not reset these statuses: the job is left in an 'executing'
state even though it is not executing, and the worker keeps the job in its map
even though it never received it.

at this point the worker pool will not assign any more jobs to that worker.

- catches errors from 'worker.send' and sets the job status to 'error'
- moves assignment of the job to the worker's data structures to 'job.execute'
  and only assigns the job if it was successfully sent to the worker thread.
- removes 'worker.execute' since it only calls 'job.execute'

* updates test of job execute