chore: Upgrade napi to latest version
We've been using a fairly old napi-rs version for quite a while due to
bugs in some of the newer versions. Try upgrading to the latest version
and see if we run into the same issues.

This would unblock the GUI app from upgrading to a newer version of
Electron, so we can officially sunset NodeJS 16 support, since 16 is
going to be entering maintenance mode soon and we want to stay current
with the LTS releases.