Verified assets: use a retry strategy when refreshing
If fetching the assets fails (because e.g. of a network glitch), do not
wait until the next refresh interval, but retry sooner. The retry
strategy used here is: exponential backoff with an initial delay of 1
minute and a random jitter of 20%.

This commit also upgrades to the version of jest, which is needed to use
`jest.advanceTimersToNextTimerAsync()`.