Add encoder name to DecoderFailed
I think it makes sense for the DecoderFailed error to contain the name
of the decoder that it was produced from. It also makes the code a bit
simpler when you are relying on accumulating errors, but need a new
wrapper type because you lose the encoder name in the error.