adds transactions to mempool in broadcastTransaction rpc
the chain/broadcastTransaction RPC endpoint can be used to prompt a node to
broadcast a valid posted transaction over the network.

however, the endpoint does not add the transaction to the node's mempool before
broadcasting. the transaction should be added to the mempool for parity with the
wallet's broadcast implementation.