verifies transaction spends in chain/broadcastTransaction (#4083)
the chain/broadcastTransaction RPC calls verifyCreatedTransaction, but that
verification only ensures that a transaction is not too large and that it has a
positive fee (current consensus rules). it does not check whether the spends in
a transaction are double spends or if the spends are valid

the wallet currently verifies spends before rebroadcasting a transaction. adding
this verification to the broadcastTransaction RPC will enable us to move that
verification from the wallet to the node