refactor(rust): remove Transaction.verify()
This function will be awkward to use since we are adding the need to
pass in mint asset owners as part of the verification process. Because
of this, and the fact that we have multiple code-paths for the same
thing, we can improve this by removing the extraneous code paths.