Fix `RawTransaction.postedSize` by considering fees when computing change notes
`RawTransaction.postedSize` contains logic to compute change notes,
which are implicit output entries that may be included in each
transaction. Currently, `postedSize` does not take into account the
transaction fee, and as such, in certain circumstances, it may
incorrectly add a change note even if such note won't be produced when
the transaction is posted.

This commit fixes the problem and adds tests that verify the output of
`RawTransaction.postedSize` with a diversified set of test transactions.