fixes error output in wallet rpc test
one of the importAccount RPC tests triggers a rescan after import. however, the
wallet database is not open in the scope of the test, so the scan throws an
error.

the test depends on the response of the RPC, which is returned before the error,
so the test succeeds. the error is thrown and outputs a stack trace.

fixes the error by mocking 'scanTransactions' for all importAccount tests to
no-op