throws error if wallet chainProcessor head not in chain (#4162)
throws an error and prevents wallet from starting instead of automatically
resetting the wallet and deleting data if the wallet's latest head hash is not
in the chain

the wallet's latest head hash might not be in the node's chain if the wallet has
synced past that node's chain head or if the wallet has synced on a fork that
the node's chain does not have. in either case the wallet cannot sync blocks on
the main chain from that node.