does not reset account data if birthday not in chain
avoids resetting account data due to network mismatch or missing forked blocks.

resets account birthday to null if the sequence is at or ahead of the head of
the node's chain and the block hash is not in the node chain. this occurs when
the birthday is on a fork or a different network.

if the birthday is on a different chain then it can't be used in scanning, so
must be set to null.

moves check of account birthdays in 'start' to follow check of chainProcessor
head so that birthdays are not reset if the chainProcessor can't scan.