stops scan for account if birthday not in chain
if we reach the sequence of an account's birthday but find that it doesn't match
the chain then the account came from a different network or fork.

we should have checked blocks before that sequence, but skipped them based on an
incorrect account birthday. starting to scan blocks for that account may lead to
incorrect wallet data.