sets createdAt to chain head for new accounts
using the head of the chain for the wallet's connected node allows the wallet to
set the account birthday based on the chain state instead of based on the state
of the other accounts in the wallet (i.e., using the chainProcessor).

sets the account head to the chain head on wallet creation.