fixes account import if wallet not conneted to node
importAccount has two dependencies on a node connection: it checks to see if the
account birthday is in the chain and it (by default) begins a scan following
import. if there is no connection to a running node the wallet cannot do either
of these things.

fixes node connection dependencies in import:
- sets account createdAt to null if it cannot be checked against the chain. this
  is what the wallet does during import if the account birthday is not found in
  the chain
- skips rescanning if there is no node connection