initializes chainProcessor in load
we recently moved the initiallization of the chainProcessor head from load to
start

this has an adverse effect for sending transactions when the walletNode is not
running: createTransaction checks whether the account "is up to date" by
comparing the account head to the chainProcessor head. if the wallet is not
started then the chainProcessor head is now null, so the account is treated as
if it were scanning.

moves chainProcessor initialization back to load so that the chainProcessor head
will not be null as long as there is an account with a non-null head when the
Wallet is loaded