Add a consensus parameter for asset ownership
Add logic for the miners fee transactions to create the correct version

mining manager only accepts transactions into blocks with the correct
version for the given sequence

add v2 support to RawTransaction (#4234)

* add v2 support to RawTransaction

* add error if transferOwnershipTo exists on a v1 transaction

check the transaction version when adding blocks

This no longer checks the version when transactions are broadcast. This
is because there isn't an easy way to verify any particular version is
correct. A transaction could become valid due to a re-org or a consensus
parameter activating at a specific block sequence.

set devnet activation to 1

fixtures

use an optimistic transaction version when creating transactions (#4237)

* use an optimistic transaction version when creating transactions

This allows for transactions created around the time that a transaction
version change to have an improved chance of being included in the
blockchain by choosing the version if its within a range near the
current head sequence instead of simply choosing the transaction version
based on head sequence + 1

* clarify `versionSequenceDelta`

- Now takes `number` instead of `number | undefined` for a parameter
- Adds an assert to ensure a non-negative integer
- Add more comments + clarify existing
- Updated tests
- Increased the min delta to 4, since 3 is probably too small to ensure
  transactions aren't dead on arrival

remove hardcoded transaction version const (#4240)

* remove hardcoded transaction version const

* add `TransactionVersion::latest()`

mempool removes transactions with versions that should no longer be valid (#4249)