fixes importAccount rescan if wallet not connected to node (#4238)
* fixes importAccount rescan if wallet not connected to node

when importing an account using ironfish-wallet-cli the wallet's 'nodeClient'
may be null.

if the nodeClient is null then it is not possible to rescan. however, that does
not mean that the wallet should use skipRescan! skipRescan sets the head of the
account equal to the head of the chainProcessor. this results in the wallet
incorrectly skipping the rescan automatically if it is not connected.

* fixes lint