When an unknown asset is seen in a wallet transaction, backfill its information from the chain db before processing the transaction
This commit is purely a refactor of the current wallet logic, and should
not alter the user-visible behavior of the wallet in any way. This
refactor is foundational work needed to support the transfer of asset
ownership (via `transferOwnershipTo`) in the future.

What this commit does, in a nutshell, is that it makes the backfilling
code run before the transaction gets processed:

    before this change: account.addPendingTransaction(...) -> backfill assets
     after this change: backfill assets -> account.addPendingTransaction(...)

This will greatly simplify the processing of mints with
`transferOwnershipTo` inside `addPendingTransaction`.