Use a floor of 1 for sequence when checking active consensus params (#4248)
Often times, this function will be called with something like
`chain.head.sequence - confirmations`. This would lead to incorrect
logic when activation sequences are set to 1 on a fresh chain.
Primarily, this will only affect devnet, but leads to simpler logic
since we won't have to special case this logic everywhere we want to
check it. A sequence will never be below 1, since the genesis sequence
is 1.