restarts wallet scan on import
when a user imports an account the wallet starts scanning the chain for
transactions involving that account.

if a user imports multiple accounts into a single wallet, then only the first
import will initiate a scan: secondary imports will not scan because a scan is
already in progress. this forces the user to restart their node to start
scanning for the other imported accounts.

updates importAccount to force a new scan to start on each import

adds 'force' parameter to 'scanTransactions'

fixes #4216