Set the default targetPeers value to 45
The default maxPeers is 50, as is the current default targetPeers. The
logic is currently such that you continuously open outbound connections
until you reach targetPeers. If targetPeers threshold is met, but
maxPeers is not, you will continue accepting inbound connections. By
setting these values separately, we create a buffer to allow for more
inbound connections. This should have a downstream effect of allowing
other peers to each peer saturation faster.