Add better peer selection logic to the syncer (#4371)
Add better peer selection logic to the syncer

The Syncer's `findPeer` function will now choose 8 random peers. Of
these peers, it will measure the round-trip-time of a
blockHeadersRequest of the genesis block. It then chooses the fastest of
these peers to sync from.

Additionally, it will check again on a growing interval (2, 4, 8... up
to 60) minutes. The purpose of this is that it should check frequently
on initial node startup, as the number of connected peers grows so that
it has a good chance of selecting a good peer to sync from early.

It includes the current syncing peer in the checks on the interval, so that
if you find a good quality peer, you are likely to continue syncing from
that peer unless you find a better one.