Use connected peers when deciding peer limits (#4350)
This uses connected nodes instead of non-disconnected nodes to determine
if a node should continue making outbound connections and accepting
inbound connections. The main change here is that we are making the
assumption that most node connections will fail, so this allows a
node to have more active connection attempts. The end goal should be
faster peer saturation. This does introduce the possibility of a
node to have more than maxPeers connected peers. The current
assumption is that natural attrition will take care of this,
however, it would be a low amount of effort to add additional logic
to prune peers if above maxPeers if it ends up being an issue.