Only loop through peer candidates if we can create new connections
Previously, this loop would get executed every event loop iteration
regardless of whether it could even create new connections. Now, it will
only run if we can actually create new connections, saving a bit of
overhead when it is not useful.

Also includes a light refactor to break this chunk of code into a
function for easier reading and testing.