Fix tests using os.tmpdir() (#4428)
This is not the right way, because os.tmpdir() does not give you a new
random directory per call. You need to create a new directory using that
as the root with a random name yourself. These tests are all affectively
using the tmpdir() root to store ironfish files in.