Change RpcServer to use RpcContext (#4441)
This allows us to isolate the requirements down to what each route
needs. This is a form of dependency injection where the route needs to
satisfy it's own dependencies, rather than have a DI system do this. A
DI system would be fine too, I just didn't want to pull one into the
code base..