Add warnings about a change to the RedDSA spec
The RedDSA specification (and the Rust crates that implement it) has
slightly deviated from the code we're using. The changes make the
signatures backward compatible, but the code we have written needs to
change in order to be compatible. This commit adds some comments to the
relevant code so that if in the future someone attempts to update the
RedDSA crates, they won't have to spend hours to figure out what's
wrong.