Add ability to extend Config
This allows you to extend Config with custom extended config types. It
does now require you to pass the extended defeults into the Config()
constructor even if you are not extending it for type reasons.

This makes the Config API slightly more burdensome to use but it's a
small price to pay.