Prompt user to enter a name for nameless account
This fix will now prompt the user to enter a name in the case that an
encoder throws an error because an account name was not provided. This
gest caught and forwarded to the front end so they can prompt the user
for a name to then try the import again.