adds memoHex to RpcWalletNote
the 'memo' included with each note is modified to remove non-human-readable
characters. if the memo included in the note is encoding data then this can make
it impossible to recover the data from the rpc response.

adds memoHex to return the memo as a hex string to preserve memo data in wallet
rpc responses