Pool: always send out a new block template (#4868)
The pool will recalculate the target of the new block template every 10 seconds. Right now, once it reaches the max target for that block, it stops sending out updated block templates with the new target and timestamp. This leads to an issue where once the max target has been reached for that block, the timestamp is never updated, so a block that takes, for example, 3 hours to mine, will still say 34 minutes because that was the last time the timestamp was updated.

By removing this check, the timestamp will be changed every 10 seconds, leading to more accurate block times for the reference pool implementation.