Refactors combine to explicitly call benchmarkSpendPostTime
Also removes `spendPostTimeAt` from internal store. In an upcoming PR this will be replaced with a different variable to create more accurate measurements for spendPostTime.
This refactor is to make the code more readable and to make it easier to understand where `spendPostTime` is being called and how the value is being calculated.