Avoid an infinite loop in `SaplingKey::generate_key`
If the random scalar generated is zero (which is a rare occurrence,
almost impossible, and has the same probability to occur as guessing
somebody's private key), `SaplingKey::generate_key` is supposed to
generate a new random scalar and try again. The current code however,
does not generate a new random scalar if that occurs, resulting in an
infinite loop.