Multisig: fetch participant identities from the public key package
When importing a multisig account, participant identities are stored in
two places:
1. the public key package (part of the account record in the walletdb)
2. on dedicated records in the walletdb

The #2 records can sometimes get lost. This is the case currently of
account rescans.

Because there is no future development scheduled that will benefit
from #2, it's easier to just ignore those records and instead rely only
on the public key package.