Move wallet/account/encoder to wallet/exporter (#4971)
* Move wallet/account/encoder to wallet/exporter

We keep running into issues where people use code in the account and
walletDB from the exporter, which they should not do. There should be no
code share between the walletDB and the exporter. This continously
causes issues where people add things to the DB and suddenly it gets
exported or expected to be decoded when someone is importing a wallet.

* Fixed lint