Move wallet scanning code into WalletScanner
This new wallet scanner tries to move some of the code (excluding
decryption) into a new class called the WalletScanner. This will
act as the holding ground for new scanning related code. This doesn't
exactly clean up a lot of the code, but does start as the beginning of
an attempt to simplify the scanning code.

For instance, the ChainProcessor is no longer exposed publically. The
wallet head now no longer returns anything meaingful in the getStatus()
RPC.