Move wallet scanning code into WalletScanner (#4983)
* Move wallet scanning code into WalletScanner

This new wallet scanner tries to move some of the code (excluding
decryption) into a new class called the WalletScanner. This will
act as the holding ground for new scanning related code. This doesn't
exactly clean up a lot of the code, but does start as the beginning of
an attempt to simplify the scanning code.

For instance, the ChainProcessor is no longer exposed publically. The
wallet head now no longer returns anything meaingful in the getStatus()
RPC.

* Fix tests

* Regenerate all fixtures

* Fixed lint

* Fixed rescan lint

* Generate slow fixtures

* Remove reset scanning during rescan route

* Fixed referencing the potentially wrong scan

* Remove wallet scan start

* Remove scan end

* Remove scan requests

* Add wallet:reset command

* WalletScanner scan always returns a ScanState

* Remove full in yup schema