avoids resetting account birthday
an account's 'createdAt' field contains the sequence and hash of the block at
which that account was created. we use it to skip decryption of notes during
rescans on blocks that came before the 'createdAt' block

we check that the 'createdAt' block is in the chain in several places to protect
against block references that may be from other networks or from pruned forks.
however, it should only be necessary to check for the account birthday when the
account is imported, and not when the wallet starts or in the middle of a scan.

by reducing the number of places that we might reset an account's 'createdAt' we
can reduce the sources of error from resetting it to null