Separate out account decoding from encrypt
This fixes the fact that we attempted to add encryption to our encoder
itself, but it was nerver fully inserted into the encoding because the
decoding flow never used the wallet to try all the secrets. This means
the encryption of an encoded account was only half in the encoding
system, and half out.

This introduces a very clear separate module at
Wallet/Exporter/Encryption which has all the methods to handle
encryption.