Improve the calculation for the number of workers in the worker pool
This commit changes the default value for `nodeWorkersMax` to the number
of CPU cores, minus one, capped at 16. Before this change,
`nodeWorkersMax` was hardcoded to 6, which resulted in limited
performance on systems that have more than 6 cores.

As a consequence of this change, most users with empty configuration
will observe a larger number of workers spawned. Users who run on
systems with SMT enabled and with a low number of cores may observe a
lower number of workers.

For example: on a system with 2 cores, with SMT enabled (4 logical CPUs
in total), and empty configuration: before this change, the number of
workers would have been 4 - 1 = 3, after this change it is 2 - 1 = 1.

The rationale behind using the number of CPU cores rather than the
number of logical CPUs is that the worker pool should mostly be used for
CPU-bound and CPU-intensive workloads. For this kind of workloads, SMT
is known to be detrimetrial for performance (other workloads that are
primarily I/O-bound should not use the worker pool but would benefit
from async code).

This commit also changes the default value for `nodeWorkers` to be based
on the number of processing units available to the process, rather than
the number of logical CPUs. On most systems, this will not result in any
difference. This change will only impact users that limit the Node
process through sandboxing tecniques like CPU affinity masks or cgroups.

Some examples to understand the impact of this change:

- users with `nodeWorkers` set in their configuration: no change
- users with `nodeWorkers`/`nodeWorkersMax` not set, 2 CPU cores, SMT
  enabled (4 CPU threads):
  number of workers before this change = min(4 - 1, 6) = 3
  number of workers after this change = min(4 - 1, 2 - 1, 16) = 1
- users with `nodeWorkers`/`nodeWorkersMax` not set, 8 CPU cores, SMT
  enabled (16 CPU threads):
  number of workers before this change = min(16 - 1, 6) = 6
  number of workers after this change = min(16 - 1, 8 - 1, 16) = 7
- users with `nodeWorkers`/`nodeWorkersMax` not set, 8 CPU cores, SMT
  disabled (8 CPU threads):
  number of workers before this change = min(16 - 1, 6) = 6
  number of workers after this change = min(8 - 1, 8 - 1, 16) = 7
- users with `nodeWorkers`/`nodeWorkersMax` not set, 32 CPU cores, SMT
  enabled (64 CPU threads):
  number of workers before this change = min(64 - 1, 6) = 6
  number of workers after this change = min(64 - 1, 32 - 1, 16) = 16
- users with `nodeWorkers`/`nodeWorkersMax` not set, 32 CPU cores, SMT
  enabled (64 CPU threads), scheduler affinity set to 4 CPUs:
  number of workers before this change = min(64 - 1, 6) = 6
  number of workers after this change = min(4 - 1, 32 - 1, 16) = 3